+++ cypress/e2e/pages/ProductListPage.js
@@ -1,0 +1,11 @@
+class Products
+    {
+
+    logo() { return cy.get(".app_logo"); }
+    cartBtn() { return cy.get(".shopping_cart_link"); }
+    pageTitle() { return cy.get(".title"); }
+
+ 
+ 
+    }
+    export default Products;
