--- cypress/e2e/pages/ProductsPage.js
@@ -1,8 +1,0 @@
-class Products
-    {
-
-    pageTitle() { return cy.get(".app_logo"); }
- 
- 
-    }
-    export default Products;
